requirements.txt
alembic==1.18.4
annotated-types==0.7.0
anyio==4.12.1
APScheduler==3.11.2
sentry-sdk[flask]==2.22.0
bcrypt==5.0.0
beautifulsoup4==4.14.3
blinker==1.9.0
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.4
distro==1.9.0
dnspython==2.8.0
email-validator==2.3.0
Flask==3.1.2
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
gunicorn==25.0.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
invoke==2.2.1
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.13.0
lxml==6.0.2
MarkupSafe==3.0.3
openai==2.16.0
packaging==26.0
tiktoken>=0.5.1
paramiko==4.0.0
psycopg2-binary==2.9.11
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyMuPDF==1.26.7
PyNaCl==1.6.2
pyotp==2.9.0
PyPDF2==3.0.1
python-docx==1.2.0
python-dotenv==1.2.1
python-http-client==3.3.7
pytz==2025.2
requests==2.32.5
sendgrid==6.12.5
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.46
tqdm==4.67.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.6.3
Werkzeug==3.1.5
WTForms==3.2.1
langchain-anthropic>=0.3.0

# Testing dependencies
pytest==8.3.5
pytest-flask==1.3.0
pytest-cov==6.0.0
alembic
bcrypt
beautifulsoup4
Flask-WTF
httpx
invoke
langchain-anthropic
pyotp
pytest
pytest-cov
pytest-flask
python-dotenv
sentry-sdk[flask]
tiktoken
tqdm
tzlocal
